fix: address CodeRabbit/Copilot review comments and CI docstring coverage
- Add docstrings to all new public items (CompletedWorkoutPowerCurve,
  PowerCurveError, compute_power_curve, PowerCurveCompletedWorkoutRepository,
  SelectedWorkoutPowerCurve, set_power_curve_5s_if_missing)
- Mongooo set_power_curve_5s_if_missing: cover null via $or filter,
  add skip_serializing_if = Option::is_none on document field,
  simplify BSON serialization inside async move closure
- power_curve: match stream_type case-insensitively,
  fall back to secondary_series when primary_series is missing
- selected_workout_power_curve tool: log persist errors with tracing::warn,
  use conditional source label (computed_5s_not_persisted on failure),
  hoist resolution validation before repository I/O
- tests: add recording to TestDataPort persist stub,
  assert persist was called in computes_5s_ad_hoc_when_cache_missing,
  extract power_curve and power_curve_repo tests to sibling files
